requirements.txt
asgiref==3.3.1
astroid==2.5.1
autopep8==1.5.6
cli-tables==0.0.2
colorama==0.4.4
Django==3.1.7
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-js-asset==1.2.2
django-modeltranslation==0.16.2
djangorestframework==3.12.4
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow==8.2.0
pycodestyle==2.7.0
pylint==2.7.2
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
